requirements_backup.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
click-log==0.4.0
colorama==0.4.6
Crawl4AI==0.7.4
cryptography==46.0.4
distro==1.9.0
exceptiongroup==1.3.1
fake-http-header==0.3.5
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.2
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanize==4.13.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.3.1
litellm==1.81.5
lxml==5.4.0
macholib @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
musicbrainzngs==0.7.1
mutagen==1.47.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
openai==2.16.0
outcome==1.3.0.post0
packaging==25.0
patchright==1.57.2
pillow==11.3.0
playwright==1.57.0
playwright-stealth==2.0.1
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyacoustid==1.3.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyOpenSSL==25.3.0
pyppeteer==2.0.0
PySocks==1.7.1
python-dotenv==1.2.1
PyYAML==6.0.3
rank-bm25==0.2.2
redis==7.0.1
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==14.3.1
rpds-py==0.27.1
rtree==1.4.1
scipy==1.13.1
selenium==4.36.0
shapely==2.0.7
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/4~CDlvugAeMYaHcTWIELyAQ2joHAecmlI0GAlKPsw/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.8.1
spotipy==2.25.2
SQLAlchemy==2.0.45
tf-playwright-stealth==1.2.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
trimesh==4.11.1
trio==0.31.0
trio-websocket==0.12.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==10.4
Werkzeug==3.1.4
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.10.14
zipp==3.23.0
